// src/app/Components/clientonboarding.enum.ts
export  enum dynamicText {
    nameText = 'Eden',
    imgName = '../../assets/logo.JPG',
    required =  'This field is required',
    firstPageHeaderTxt = 'Welcome! First things first...',
    firstPageText = 'You can always change them later.',
    fullName = 'Full Name',
    displayName = 'Display Name',
    createBtn = 'Create Workspace',
    setupHeaderTxt = 'Let`s setup a home for all your work',
    setupMsgTxt = 'You can always create another workspace later.',
    workspaceName = 'Workspace Name',
    workspaceUrl = 'Workspace URL',
    optional = '(optional)',
    setupTitle = 'How are you planning to use Eden?',
    planText = 'We`ll stremline your setup accordingly.',
    congratsImg = '../../assets/congarts.JPG', 
    congratsTitle = 'Congratulations, Eren!',
    launchMsg = 'You have completed onboarding, you can start Eden!',
    launchBtn =  'Launch Eden',
}